Move token and error into related modules
