+++ crates/ruff_python_parser/src/token_source.rs
@@ -1,7 +1,9 @@
 use ruff_text_size::{Ranged, TextRange, TextSize};
 
-use crate::lexer::{Lexer, LexerCheckpoint, LexicalError, Token, TokenFlags, TokenValue};
-use crate::{Mode, TokenKind};
+use crate::error::LexicalError;
+use crate::lexer::{Lexer, LexerCheckpoint};
+use crate::token::{Token, TokenFlags, TokenKind, TokenValue};
+use crate::Mode;
 
 /// Token source for the parser that skips over any trivia tokens.
 #[derive(Debug)]
